Update compiler hook to webpack 5 API
The compiler.hooks.compilation.tap method is used to attach a new hook to the compilation phase of the Webpack build process. In this case, the hook is called "Rename.emit" and the callback function that is passed to it is executed when the compilation phase starts.

The callback function uses the compilation.hooks.optimizeAssets.tapAsync method to attach another hook to the optimizeAssets phase of the compilation process. This hook is called "Rename.emit" and it's a async hook, which means it takes a callback function that should be called when the hook is finished.

The callback function takes two arguments: assets, the assets object, and callback which is a function that should be called when the hook is finished.